Refactor timers
